Fixes for latest clang
 * Don't auto-include the pch while compiling the pch.
 * Fix several bugs found by clang (comparisons that were always true)

Reviewed-on: https://gerrit.dechocorp.com/10555

+++ mordor/http/multipart.cpp
@@ -132,9 +132,9 @@ public:
 
     size_t read(Buffer &b, size_t len)
     {
-        size_t boundary = parent()->find(m_boundary, len, false);
+        ptrdiff_t boundary = parent()->find(m_boundary, len, false);
         if (boundary >= 0)
-            len = std::min(boundary, len);
+            len = std::min((size_t)boundary, len);
         return parent()->read(b, len);
     }
